Move UC3 tests out of subfolders
